requirements.txt
discord.py-self==2.0.0
openai==0.28.1
tiktoken>=0.3.0
python-dotenv>=0.20.0
asyncio==3.4.3
requests>=2.28.0